requirements.txt
# Core dependencies
python-dotenv>=1.0.0
tabulate>=0.9.0
matplotlib>=3.7.1
pandas>=2.0.0
numpy>=1.24.0

# AutoGen dependencies
pyautogen>=0.2.0

# Semantic Kernel dependencies
semantic-kernel>=0.3.0

# LangChain dependencies
langchain>=0.0.267
langchain-experimental>=0.0.42
langgraph>=0.0.17

# CrewAI dependencies
crewai>=0.1.0

# LLM providers
openai>=1.1.0
groq>=0.4.0

# Optional: Vector database for some examples
chromadb>=0.4.0 